:construction: Rework AIG adapters to their own translation units
